move: 중첩된 message response dto 폴더 제거
